setting: 회원가입 fsd 구조로 쪼개기
