+++ src/shared/lib/assets.ts
@@ -1,0 +1,14 @@
+export const collageItems = [
+  'ICT융합대학',
+  '경영대학',
+  '사회과학대학',
+  '인문대학',
+  '법과대학',
+  '미래융합대학',
+  '자연과학대학',
+  '공과대학',
+  '예술체육대학',
+  '건축대학',
+  '방목기초교육대학',
+];
+export const departmentItems = ['융합소프트웨어학부', '법학과', '경영학과'];
